requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
langgraph==0.2.25
langchain-core==0.3.13
langchain-openai==0.2.3
